Changed the code from line count to word count
One thing to note is that perviously, when I was counting the # of lines, I was creating an RDD forcefully so that I would be able to write an output file.
That was because, when we check the # of lines the output of the function returns an integer, it does not return an RDD, so I had to use parallelise to get an RDD out to create an output file.

Whereas while I am doing the word count now, naturally the function returns an RDD, so no additional steps required.

I am told, previously when the code was trying to count the # of lines, I would have had to use S3 api's to be able to write output to a file